+++ requirements.txt
@@ -2,4 +2,4 @@ pycryptodome~=3.20.0
 crypto~=1.4.1
 requests~=2.31.0
 wget~=3.2
-html2docx~=1.6.0
+xhtml2pdf~=0.2.16
